bhyve: Move some more amd64-specific drivers to their own subdir
No functional change intended.

Reviewed by:	corvink, jhb, emaste
MFC after:	1 week
Sponsored by:	Innovate UK
Differential Revision:	https://reviews.freebsd.org/D40551

(cherry picked from commit f927afc1a6fe8257e28d79e8bb77305c2958724a)
