+++ usr.sbin/bhyve/amd64/ps2mouse.c
